fix: ai depth beyond 1 card (recursion)
